+++ Sources/CircularArray/CircularArray.swift
@@ -7,7 +7,7 @@
 //
 
 import Destructure
-import Restructure
+import Algebra
 
 /// Array-like structure that allows retrieval of elements at indices outside of the bounds of
 /// the internal storage.
