+++ Sources/ReferenceTree/ReferenceTreeProtocol.swift
@@ -6,7 +6,7 @@
 //
 //
 
-import Restructure
+import Algebra
 
 /// Error for when you do bad things to a `ReferenceTreeProtocol`-conforming type.
 public enum ReferenceTreeError: Error {
